Reconnect expected audio devices dynamically

+++ recs/ui/source_process.py
@@ -22,12 +22,18 @@ class SourceProcess(Runnable):
         self.name = tracks[0].source.name
         self.source = tracks[0].source
         self.tracks = tracks
+        self.started: bool = False
+
+    @property
+    def required_channels(self) -> int:
+        return max(int(channel) for track in self.tracks for channel in track.channels)
 
     @property
     def is_alive(self) -> bool:
-        return self.process.is_alive()
+        return self.started and self.process.is_alive()
 
     def start(self) -> None:
+        assert not self.started
         self.connection, child = mp.Pipe()
         self.stop_event = mp.Event()
         kwargs = {
@@ -38,10 +44,13 @@ class SourceProcess(Runnable):
         }
         self.process = mp.Process(target=SourceRecorder, kwargs=kwargs)
         self.process.start()
+        self.started = True
         self.stopped = False
         super().start()
 
     def stop(self) -> None:
+        if not self.started:
+            return
         self.running = False
         self.stop_event.set()
 
@@ -49,9 +58,13 @@ class SourceProcess(Runnable):
         self.stop()
 
     def join(self, timeout: float | None = None) -> None:
+        if not self.started:
+            return
         self.process.join(STOP_TIMEOUT if timeout is None else timeout)
         if self.process.is_alive():
             self.process.terminate()
             self.process.join()
         self.connection.close()
+        self.running = False
+        self.started = False
         self.stopped = True
